chore(deps): Bump aws-sdk to v2.946.0
We pinned this dependency in #25 because of issues observed in some builds that ran `npx @guardian/node-riffraff-artifact`.

It seems the root cause of this the node Docker image using the root user as default.

As we know the root cause and have a recommendation against using `npx` in CI as it is non-deterministic,
we can bump `aws-sdk` normal now.

See:
  - https://github.com/aws/aws-sdk-js/issues/3763
  - https://github.com/guardian/recommendations/pull/60
  - https://github.com/guardian/node-riffraff-artifact/pull/25